CI: don't generate GRAPHVIZ_VERSION file in macOS CMake build job
The version is generated internally to a variable in the CMake build
job and the downstream test job generates the file itself since commit
fea17b6a9738eb96f037e22538d1885a75f20698.

This commit moves the generation of the GRAPHVIZ_VERSION file from the
build template to the macOS autotools build job.