gvconfig.c: realign alternative for 'glob' with POSIX
If the `glob` system function¹ is available, it is used in this code. If not (on
Windows), an alternative static function is provided. The POSIX spec mandates
only three members of the `glob_t` struct, `gl_pathc`, `gl_pathv`, and
`gl_offs`. This commit realigns the members with their types in the POSIX spec.
This squashes some -Wsign-compare warnings on POSIX platforms.

¹ https://pubs.opengroup.org/onlinepubs/007908799/xsh/glob.html