remove 'COLLECTION' from the packaging hierarchy
Each Graphviz version has a notion of “collection,” which traditionally could be
the values “stable” or “development.” The semantically versioned releases that
are published to the website are considered stable. The inter-release packages
that the deploy.py script still uploads to the Gitlab package repository are
considered development. gen_version.py knows how to determine the collection of
a version during build.

Additionally .gitlab-ci.yml knew of another collection value, “experimental,” of
which gen_version.py was unaware. It would manually tweak the collection to this
for certain non-mainstream builds.

deploy.py knows nothing of this notion of collections.

The combined effect of the above has led to some negative outcomes:

  1. Published releases include a mixture of files labelled “stable” and files
     labelled “experimental.” It is not clear to the general public what either
     of these labels mean or how they differ.

  2. In recent times, “development” versions are no more or less stable than
     “stable” versions. The terminology here unfortunately has become
     inaccurate.

  3. The directory hierarchy for published releases has included an extra level
     for collection. If deployment worked as intended, this would be redundant
     as every release artifact would be “stable.” However, as discussed above,
     the deployment script actually publishes “experimental” artifacts as well.

To reduce confusion and simplify build and packaging going forwards, this change
removes the notion of collection. The generated COLLECTION file is no longer
produced and users should have no need to be aware of this concept.
gen_version.py still internally knows the idea of collections to distinguish
published release version numbers from inter-release version numbers. But this
is no longer exposed to the rest of the build system and packaging.

The notion of “experimental” as a collection value is completely removed.

Closes #2148.