fix: use 'find_program' instead of 'find_package' to locate Git in CMake build
The Cygwin CMake CI task recently started stalling after the “Found FLEX” step.
Reordering CMake steps to debug this revealed that it was definitely the
`find_package(Git REQUIRED)` step that was stalling. I have not been able to
root cause what is going on, but some educated guesses:

  1. A CMake FindGit.cmake bug? This could certainly cause the described
     behavior, but I cannot find anything online to corroborate this and cannot
     see anything in the upstream history of this file¹ that looks like a fix
     for something like this.

  2. Installing a version of CMake via Choco and then running it via Cygwin
     causes some bad Windows/UNIX interaction? Perhaps, but I cannot find any
     corroborating evidence for this either. Also, I cannot guess why such a
     thing would _only_ cause problems in FindGit.cmake.

  3. FindGit.cmake is picking up a Visual Studio shim for Git that responds in a
     way that confuses the `find_package` check? I found vague hints about a
     shim like this online, but I still cannot explain why this would cause a
     stall.

A work around for this is to use the less nuanced `find_program`. This has no
negative impact on the CMake setup because (1) all we are attempting to detect
is the `git` binary itself and (2) the CMake files never call `git` directly
anyway; we are performing this detection to avoid gen_version.py failing
cryptically.

Note that it is not possible to use `REQUIRED` in this `find_program` directive
because this was only added in CMake 3.18 and we currently only require CMake
3.1.

This also re-enables the Cygwin CMake CI task now that it no longer stalls.

Fixes #2145.

¹ https://github.com/Kitware/CMake/commits/master/Modules/FindGit.cmake