gvrender_beziercurve: realign 'filled' parameter type
This appears to have been a mistake, in that all the other `gvrender_*`
functions take an `int` for the `filled` parameter. This does not appear to have
caused any visible effects, as the values passed to this function always fit in
a `boolean`. But the compiler complained that it thought unintended value
truncation was occurring.

Squashes 3 -Wconversion compiler warnings.