further consolidate RPMs into a single tarball per OS
Taking the previous approach further, this results in the following release
artifacts being combined:

  * CentOS 7
    * 1 source RPM, 2 tarballs → 1 tarball
  * CentOS 8
    * 1 source RPM, 2 tarballs → 1 tarball
  * Fedora 33
    * 1 source RPM, 2 tarballs → 1 tarball
  * Fedora 34
    * 1 source RPM, 2 tarballs → 1 tarball

This drops the total number of Graphviz release artifacts from 176 to 168. This
is still not below the 50 artifact ceiling we need to hit, but this is progress.

Note that this removes the `os`/`debug`/`source` hierarchy in the packaging
directories for these artifacts, now that all three of these are packaged into a
single tarball. The `os` hierarchy remains for the CMake-produced RPM for these
platforms.

Related to #2147.