further consolidate DEBs into a single tarball per OS
Taking the previous approach further, this results in the following release
artifacts being combined:

  * Ubuntu 18.04
    * 13 DEBs, 10 debug DEBs → 1 tarball
  * Ubuntu 20.04
    * 13 DEBs, 10 debug DEBs → 1 tarball
  * Ubuntu 20.04 debug
    * 13 DEBs, 10 debug DEBs → 1 tarball
  * Ubuntu 21.04
    * 13 DEBs, 10 debug DEBs → 1 tarball
  * Ubuntu 21.10
    * 13 DEBs, 10 debug DEBs → 1 tarball

This drops the total number of Graphviz release artifacts from 63 to 58. This is
still not below the 50 artifact ceiling we need to hit, but this is progress.

Note that this removes the `os`/`debug`/`source` hierarchy in the packaging
directories for these artifacts, now that all three of these are packaged into a
single tarball. The `os` hierarchy remains for the CMake-produced DEB for these
platforms.

Related to #2147.