general.h: stop disabling assertions when 'DEBUG' is undefined
The effect of this code was that #including general.h prior to assert.h would
disable assertions in your code. Not #including general.h or #including it after
assert.h would allow assertions to be controlled through the standard `NDEBUG`
mechanism. It is not clear why suppressing assertions like this was ever
desirable.

This is basically another variant of commit
68bcbbd4ddef4960c7ea6884dda10854b0e012e0.