remove 'real' abstraction
This was not configurable and always mapped to the type `double`. There were
also implicit assumptions that `real` and `double` were interchangeable. This
alias is removed to avoid future confusion.

The diff appears large, but it is just removing two #defines of `real` and
propagating their definition (`double`) to all usage sites.