[nfc] remove legacy 'STACK_DIRECTION' lines in Windows config.h
This constant was introduced in the initial addition of the precreated Windows
config.h, 76ead5c74029a5ceea61dd6c7600dc1931c275cf, but it is unclear where it
originates from. None of the other build systems appear to define this and no
prior or current state of the Graphviz tree has ever contained use of this
macro. Graphviz does not use `alloca`.