makeStraightEdge: remove array allocation optimization
This code had two alternative paths, one that used a heap-allocated array for
edge counts above 20 and one that used a stack-allocated array for edge counts
of 20 or less. The latter was an optimization to avoid heap allocation for
smaller graphs. Modern dynamic allocators are fast enough that this kind of
optimization is no longer relevant.