common: include utils.h and types.h in timing.c also for _WIN32
This ensures that the timing functions are declared with the
storage-class attributes from utils.h.

An upcoming commit that corrects the storage-class attribute
declarations in utils.h would otherwise cause these errors:

/usr/lib/gcc/x86_64-pc-cygwin/10/../../../../x86_64-pc-cygwin/bin/ld: ../../lib/dotgen/.libs/libdotgen_C.a(mincross.o): in function `init_mincross':
/home/magja/graphviz/lib/dotgen/mincross.c:1163: undefined reference to `__imp_start_timer'
/usr/lib/gcc/x86_64-pc-cygwin/10/../../../../x86_64-pc-cygwin/bin/ld: ../../lib/dotgen/.libs/libdotgen_C.a(mincross.o): in function `cleanup2':
/home/magja/graphviz/lib/dotgen/mincross.c:996: undefined reference to `__imp_elapsed_sec'

The reason for also including types.h is that utils.h depends on it
without including it itself for unclear reasons.

Note that these functions are also redundantly declared in
timing.h. This file will be removed in an upcoming commit series.

Towards https://gitlab.com/graphviz/graphviz/-/issues/2058.