remove cgraph macro aliases
The functions removed in this commit were intended to be aliases for the
same-named macros declared in cgraph.h. The claimed purpose is debugger
introspection.

It is unclear to me what value these have in a modern development environment.
Some modern debuggers can understand macro definitions. Those that cannot are
_impeded_ by these aliases. A developer debugging looks at the alias definition
and thinks they understand the code, but they do not. The code that is executing
on their system is _not_ the code these functions compile to. To state something
stronger: it is unclear to me how these definitions were ever helpful. They
appear to (accidentally) obscure what is actually occurring in the machine at
runtime.

Note that these functions also bloated the cgraph .so/.a size by including
functions that were never called.