API BREAK: remove local 'INT_MAX' and 'INT_MIN' definitions
These constants are defined in ISO C to be available in limits.h, but were
commonly available pre-ISO C too. These definitions remained here for legacy
reasons, because removing them was considered an API break and not worth doing
for this minor clean up. However the upcoming release is projected to involve a
major version number bump.