API BREAK: remove unused 'MAXSHORT' definition
The last use of this was removed in 580fa88477fa59aca01e405c986c686ceee21b5f.
But its definition was retained in this header as removing it was considered an
API break not worth causing. The upcoming release is projected to bump the major
version number, so we can now remove it.