Revert "sed cleanup of CFLAGS from perl which break on opencsw - should be harmless elsewhere"
This reverts commit a750d28463a9248fe5a42d3dae5768d38853d837. It is unclear to
me exactly what flag the original code was attempting to suppress, but this
regex is too broad. The specific case causing problems right now is that it
incorrectly mangled `-mno-xop -mbmi` into `-mno-mbmi`. Given the original code
appears to have been working around a 2011 OpenCSW packaging mistake, lets
assume this was fixed sometime in the last decade and see if anyone downstream
complains when we remove the work around.

Gitlab: fixes #2177