mingle bundle: replace a 'unique_ptr' with a 'vector'
Similar to the prior commit, when the `std::unique_ptr` usage was added here in
176c8a163a951a2b801956ed4a00da05603e3a06, I was incorrectly too focused on
preserving the previous property of this being a heap-allocated array. This is
not necessary, and using a `std::vector` instead allows more flexibility. E.g.
depending on the `std::vector` implementation, it could choose to allocate this
array on the stack instead of the heap.