CI: add Automake installation for macOS
Automake (and its dependent, Autoconf) are needed for Autotools-based Graphviz
compilation. They are currently available in the CI Gitlab macOS 11 shared
runner environment incidentally due to a transitive dependency via php.¹ The
same is not true in the macOS 12 environment.²

This commit moves to explicitly installing these Graphviz dependencies rather
relying on getting them implicitly. This makes the CI task for the macOS
Autotools build pass on macOS 12.

Gitlab: #2164

¹ https://gitlab.com/gitlab-org/ci-cd/shared-runners/images/macstadium/orka/-/blob/df5b9cad3a482ec8b79b401c8d03bbc60f2f97c4/toolchain/big-sur.yml#L105-111
² https://gitlab.com/gitlab-org/ci-cd/shared-runners/images/macstadium/orka/-/blob/df5b9cad3a482ec8b79b401c8d03bbc60f2f97c4/toolchain/monterey.yml#L48-49