CMake: add a variable substitution step for plugin configuration step
This CMake script hard codes some library paths in a way that thwarts
user/system preferences (#1973). Attempts to solve this by falling back on
standard CMake mechanisms have been unsuccessful because an install script like
this runs in an isolated environment where it (correctly) does not have access
to the configuration from the build environment. To resolve this we need a way
to pass the install script parameters/arguments from the build context.
Introducing a level of indirection here gives us such a mechanism.

This change by itself does nothing. However an upcoming change will take
advantage of the `@…@` substitution mechanism enabled by this change.

Note that the word “configure” in these files is used to refer to two distinct
mechanisms:

  1. The CMake `configure_file` command does something analogous to Autotools’
     variable substitution during `./configure`.

  2. The Graphviz `dot -c …` configure step registers installed Graphviz
     plugins.

Gitlab: #1973